Price Column added to Product domain model

+++ IydeParfume/IydeParfume/Areas/Admin/ViewModels/Product/ListItemViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class ListItemViewModel
     {
-        public ListItemViewModel(int id, string name, string description, int price, string imageUrl)
+        public ListItemViewModel(int id, string name, string description, decimal price, string imageUrl)
         {
             Id = id;
             Name = name;
@@ -14,7 +14,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string ImageUrl { get; set; }
     }
 }
